[FEAT] TeamSp API

+++ src/main/java/SSOP/ssop/service/UserService.java
@@ -5,13 +5,12 @@ import SSOP.ssop.domain.User;
 import SSOP.ssop.dto.LoginDto;
 import SSOP.ssop.dto.UserDto;
 import SSOP.ssop.repository.UserRepository;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
@@ -70,28 +69,72 @@ public class UserService {
     }
 
     // 특정 유저 정보 출력
-    public Map<String, Object> getUser(long userId) {
+    public ResponseEntity<?> getUser(long userId) {
         Optional<User> user = userRepository.findByUserId(userId);
         if (user.isPresent()) {
-            return Collections.singletonMap("user", new UserDto(user.get()));
+            // UserDto 객체를 직접 반환
+            return ResponseEntity.ok(new UserDto(user.get()));
         } else {
-            return Collections.singletonMap("message", "존재하지 않는 사용자입니다.");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(Map.of("message", "존재하지 않는 사용자입니다."));
         }
     }
 
-    // 유저 정보 업데이트 -> 일단 pw만
-    public void updateUser(UserDto userDto) {
-        User user = userRepository.findById(userDto.getUserId())
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
+    // 유저 비밀번호 수정
+    public ResponseEntity<?> updatePassword(UserDto userDto) {
+        if (!userRepository.existsById(userDto.getUserId())) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(Map.of("message", "존재하지 않는 사용자입니다."));
+        }
+
+        User user = userRepository.findById(userDto.getUserId()).get();
         user.setPassword(passwordEncoder.encode(userDto.getPassword())); // 비밀번호 암호화
         userRepository.save(user);
+
+        return ResponseEntity.ok().body(Map.of("message", "비밀번호가 성공적으로 변경되었습니다."));
+    }
+
+    // 유저 전화번호 수정
+    public ResponseEntity<?> updatePhone(UserDto userDto) {
+        if (!userRepository.existsById(userDto.getUserId())) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(Map.of("message", "존재하지 않는 사용자입니다."));
+        }
+
+        User user = userRepository.findById(userDto.getUserId()).get();
+        user.setUser_phone(userDto.getUser_phone());
+        userRepository.save(user);
+
+        return ResponseEntity.ok().body(Map.of("message", "전화번호가 성공적으로 변경되었습니다."));
+    }
+
+    // 유저 이름 & 생년월일 수정
+    public ResponseEntity<?> updateNameBirth(UserDto userDto) {
+        if (!userRepository.existsById(userDto.getUserId())) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(Map.of("message", "존재하지 않는 사용자입니다."));
+        }
+
+        User user = userRepository.findById(userDto.getUserId()).get();
+        user.setUser_name(userDto.getUser_name());
+        user.setUser_birth(userDto.getUser_birth());
+        userRepository.save(user);
+
+        return ResponseEntity.ok().body(Map.of("message", "회원 정보가 성공적으로 변경되었습니다."));
     }
 
     // userId로 유저 삭제
-    public void deleteUser(long userId) {
-        User user = userRepository.findByUserId(userId)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
+    public ResponseEntity<Map<String, String>> deleteUser(long userId) {
+        Optional<User> userOptional = userRepository.findByUserId(userId);
 
-        userRepository.delete(user);
+        if (userOptional.isPresent()) {
+            User user = userOptional.get();
+            userRepository.delete(user);
+            return ResponseEntity.ok(Map.of("message", "탈퇴되었습니다.."));
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(Map.of("message", "존재하지 않는 사용자입니다."));
+        }
     }
+
 }
